pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opstwin-recovery"
version = "1.0.0"
description = "OpsTwin Recovery Arena -- enterprise incident response OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
train = [
    "trl>=0.14",
    "transformers>=4.45",
    "peft>=0.13",
    "datasets>=2.18",
    "accelerate>=0.30",
    "vllm",
    "trackio",
]
inference = [
    "openai>=1.0.0",
    "python-dotenv>=1.0",
]
eval = [
    "matplotlib>=3.8",
    "torch",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["server"]
py-modules = ["models", "client", "inference", "train", "evaluate"]

[tool.setuptools.package-data]
server = ["*.py"]
